backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
torch==2.2.0
numpy==1.26.4
gunicorn==21.2.0
whitenoise==6.6.0
